Delete question

+++ FoodApp/FoodApp/ShopinglistPage.xaml.cs
@@ -51,12 +51,18 @@ namespace FoodApp
 
         private async void RemoveItemButton_Clicked(object sender, EventArgs e)
         {
-            var item = (Button)sender;
-            var itemId = item.ClassId;
-            await apiService.DeleteItemAsync(itemId);
-            App.Current.MainPage = new ShopinglistPage();
-        }
+            var item = (ImageButton)sender;
+            bool answer = await DisplayAlert("Delete", "Would you like to remove the item?", "Yes", "No");
+            if(!answer)
+            {
 
+            }
+            else { 
+                var itemId = item.ClassId;
+                await apiService.DeleteItemAsync(itemId);
+                App.Current.MainPage = new ShopinglistPage();
+            }
+        }
         private async void AddItemButton_Clicked(object sender, EventArgs e)
         {
             string result = await DisplayPromptAsync("Add", "What do you whant to add?");
